Rename to testReposIssues(), move other tests over
